move flags
