+++ cmd/process.go
@@ -3,20 +3,21 @@ package cmd
 import (
 	"context"
 	"fmt"
+	"net/http"
 	"os"
+	"strconv"
 	"time"
 
 	"cosmossdk.io/log"
+	"github.com/gin-gonic/gin"
+	"github.com/rs/zerolog"
 	"github.com/spf13/cobra"
 	"github.com/strangelove-ventures/noble-cctp-relayer/circle"
+	"github.com/strangelove-ventures/noble-cctp-relayer/ethereum"
+	"github.com/strangelove-ventures/noble-cctp-relayer/noble"
 	"github.com/strangelove-ventures/noble-cctp-relayer/types"
+	"gopkg.in/yaml.v2"
 )
-
-var startCmd = &cobra.Command{
-	Use:   "start",
-	Short: "Start relaying CCTP transactions from Ethereum to Noble",
-	Run:   Start,
-}
 
 // State and Store map the iris api lookup id -> MessageState
 // State represents all in progress burns/mints
@@ -26,41 +27,50 @@ var State = types.NewStateMap()
 // SequenceMap maps the domain -> the equivalent minter account sequence or nonce
 var sequenceMap = types.NewSequenceMap()
 
-func Start(cmd *cobra.Command, args []string) {
+func Start() *cobra.Command {
+	cmd := &cobra.Command{
+		Use:   "start",
+		Short: "Start relaying CCTP transactions",
+		Run: func(cmd *cobra.Command, args []string) {
 
-	// messageState processing queue
-	var processingQueue = make(chan *types.TxState, 10000)
+			// messageState processing queue
+			var processingQueue = make(chan *types.TxState, 10000)
 
-	registeredDomains := make(map[types.Domain]types.Chain)
+			registeredDomains := make(map[types.Domain]types.Chain)
 
-	for name, cfg := range Cfg.Chains {
-		c, err := cfg.Chain(name)
-		if err != nil {
-			Logger.Error("Error creating chain", "err: ", err)
-			os.Exit(1)
-		}
+			for name, cfg := range Cfg.Chains {
+				c, err := cfg.Chain(name)
+				if err != nil {
+					Logger.Error("Error creating chain", "err: ", err)
+					os.Exit(1)
+				}
 
-		if err := c.InitializeBroadcaster(cmd.Context(), Logger, sequenceMap); err != nil {
-			Logger.Error("Error initializing broadcaster", "err: ", err)
-			os.Exit(1)
-		}
+				if err := c.InitializeBroadcaster(cmd.Context(), Logger, sequenceMap); err != nil {
+					Logger.Error("Error initializing broadcaster", "err: ", err)
+					os.Exit(1)
+				}
 
-		go c.StartListener(cmd.Context(), Logger, processingQueue)
+				go c.StartListener(cmd.Context(), Logger, processingQueue)
 
-		if _, ok := registeredDomains[c.Domain()]; ok {
-			Logger.Error("Duplicate domain found", "domain", c.Domain(), "name:", c.Name())
-			os.Exit(1)
-		}
+				if _, ok := registeredDomains[c.Domain()]; ok {
+					Logger.Error("Duplicate domain found", "domain", c.Domain(), "name:", c.Name())
+					os.Exit(1)
+				}
 
-		registeredDomains[c.Domain()] = c
+				registeredDomains[c.Domain()] = c
+			}
+
+			// spin up Processor worker pool
+			for i := 0; i < int(Cfg.ProcessorWorkerCount); i++ {
+				go StartProcessor(cmd.Context(), Cfg, Logger, registeredDomains, processingQueue, sequenceMap)
+			}
+
+			<-cmd.Context().Done()
+		},
 	}
-
-	// spin up Processor worker pool
-	for i := 0; i < int(Cfg.ProcessorWorkerCount); i++ {
-		go StartProcessor(cmd.Context(), Cfg, Logger, registeredDomains, processingQueue, sequenceMap)
-	}
-
-	<-cmd.Context().Done()
+	cmd.Flags().StringVarP(&cfgFile, "config", "c", "config.yaml", "file path of config file")
+	cmd.Flags().BoolVarP(&verbose, "verbose", "v", false, "use this flag to set log level to `debug`")
+	return cmd
 }
 
 // StartProcessor is the main processing pipeline.
@@ -190,5 +200,96 @@ func LookupKey(sourceTxHash string) string {
 }
 
 func init() {
-	cobra.OnInitialize(func() {})
+	cobra.OnInitialize(func() {
+		if verbose {
+			Logger = log.NewLogger(os.Stdout)
+		} else {
+			Logger = log.NewLogger(os.Stdout, log.LevelOption(zerolog.InfoLevel))
+		}
+
+		var err error
+		Cfg, err = Parse(cfgFile)
+		if err != nil {
+			Logger.Error("unable to parse config file", "location", cfgFile, "err", err)
+			os.Exit(1)
+		}
+		Logger.Info("successfully parsed config file", "location", cfgFile)
+
+		// start api server
+		go startApi()
+	})
+}
+
+func startApi() {
+	gin.SetMode(gin.ReleaseMode)
+	router := gin.Default()
+
+	err := router.SetTrustedProxies(Cfg.Api.TrustedProxies) // vpn.primary.strange.love
+	if err != nil {
+		Logger.Error("unable to set trusted proxies on API server: " + err.Error())
+		os.Exit(1)
+	}
+
+	router.GET("/tx/:txHash", getTxByHash)
+	router.Run("localhost:8000")
+}
+
+func getTxByHash(c *gin.Context) {
+	txHash := c.Param("txHash")
+
+	domain := c.Query("domain")
+	domainInt, err := strconv.ParseInt(domain, 10, 0)
+	if domain != "" && err != nil {
+		c.JSON(http.StatusBadRequest, gin.H{"message": "unable to parse domain"})
+	}
+
+	if tx, ok := State.Load(txHash); ok && domain == "" || (domain != "" && tx.Msgs[0].SourceDomain == types.Domain(domainInt)) {
+		c.JSON(http.StatusOK, tx.Msgs)
+		return
+	}
+
+	c.JSON(http.StatusNotFound, gin.H{"message": "message not found"})
+}
+
+func Parse(file string) (*types.Config, error) {
+	data, err := os.ReadFile(file)
+	if err != nil {
+		return nil, fmt.Errorf("failed to read file %w", err)
+	}
+
+	var cfg types.ConfigWrapper
+	if err := yaml.Unmarshal(data, &cfg); err != nil {
+		return nil, fmt.Errorf("error unmarshalling config: %w", err)
+	}
+
+	c := types.Config{
+		EnabledRoutes:        cfg.EnabledRoutes,
+		Circle:               cfg.Circle,
+		ProcessorWorkerCount: cfg.ProcessorWorkerCount,
+		Api:                  cfg.Api,
+		Chains:               make(map[string]types.ChainConfig),
+	}
+
+	for name, chain := range cfg.Chains {
+		yamlbz, err := yaml.Marshal(chain)
+		if err != nil {
+			return nil, err
+		}
+
+		switch name {
+		case "noble":
+			var cc noble.ChainConfig
+			if err := yaml.Unmarshal(yamlbz, &cc); err != nil {
+				return nil, err
+			}
+			c.Chains[name] = &cc
+		default:
+			var cc ethereum.ChainConfig
+			if err := yaml.Unmarshal(yamlbz, &cc); err != nil {
+				return nil, err
+			}
+			c.Chains[name] = &cc
+		}
+	}
+	return &c, err
 }
